+++ test/runtests.jl
@@ -20,4 +20,5 @@ include("fixtures/build_fixture.jl")
     include("test_mp_system_holdout.jl")
     include("test_element_pair_model.jl")
     include("test_mp_element_pair.jl")
+    include("test_scripts.jl")
 end